move a function to utlity
