+++ hft/output.py
@@ -83,30 +83,6 @@ class MarketRecord(TimeAwareInSessionRecord):
     e_best_offer = models.IntegerField()
     signed_volume = models.FloatField()
     e_signed_volume = models.FloatField()
-    
-def serialize_in_memo_model(in_memo_model, req_props, req_subprops: dict):
-    # assumes no clash in tuple elements and dict keys
-    result = {}
-    if req_props:
-        for prop_name in req_props: 
-            value = getattr(in_memo_model, prop_name)
-            if value is not None:
-                result[prop_name] = value
-    if req_subprops:
-        for prop_name, subprop_names in req_subprops.items():
-            attr = getattr(in_memo_model, prop_name)
-            if attr is not None:
-                for subprop_name in subprop_names:
-                    if isinstance(attr, dict):
-                        value = attr[subprop_name]
-                    elif hasattr(attr, subprop_name):
-                        value = getattr(attr, subprop_name)
-                        # maybe it is a dictionary
-                    else:
-                        raise AttributeError('%s has no %s:%s' % (attr, prop_name, subprop_name))
-                    result[subprop_name] = value
-    return result
-
 
 def get_required_model_fields(session_format, model_name):
     market_env = market_environments.environments[session_format].checkpoint[model_name]
@@ -115,19 +91,6 @@ def get_required_model_fields(session_format, model_name):
     return req_props, req_subprops
 
 
-# def trader_checkpoint(serialized_model: dict, session_format, event_type='', event_no=None,
-#                        market_environments=market_environments):
-#     def ensure_valid_kws(record_cls, kwas: dict):
-#         keys = list(kwas.keys())
-#         fields_names = [f.name for f in record_cls._meta.fields]
-#         for k in keys:
-#             if k not in fields_names:
-#                 del kwas[k]
-#         return kwas
-#     kws = ensure_valid_kws(InSessionTraderRecord, serialized_model)
-#     record = InSessionTraderRecord.objects.create(trigger_event_type=event_type,
-#         event_no=event_no, **kws)
-#     return record
 IN_SESSION_RECORD_CLASSES={'trader': TraderRecord, 'market': MarketRecord,
                 'inv': TraderRecord}
 def checkpoint(serialized_model: dict, session_format, model_name,
